restructure types + add test dir
