+++ spec/src/types/accounts.ts
@@ -1,0 +1,8 @@
+
+import { IdlTypeDefTyStruct } from "./definedTypes";
+
+export type IdlAccountDef = {
+    name: string;
+    docs?: string[];
+    type: IdlTypeDefTyStruct;
+};
