+++ spec/src/types/constants.ts
@@ -1,0 +1,8 @@
+
+import { IdlType } from "./type";
+
+export type IdlConstant = {
+    name: string;
+    type: IdlType;
+    value: string;
+};
