+++ spec/src/types/field.ts
@@ -1,0 +1,18 @@
+import { IdlType } from "./type";
+
+export type IdlField = {
+    name: string;
+    docs?: string[];
+    type: IdlType;
+};
+
+export type IdlEnumVariant = {
+    name: string;
+    fields?: IdlEnumFields;
+};
+
+export type IdlEnumFields = IdlEnumFieldsNamed | IdlEnumFieldsTuple;
+
+export type IdlEnumFieldsNamed = IdlField[];
+
+export type IdlEnumFieldsTuple = IdlType[];
