+++ spec/src/types/index.ts
@@ -1,0 +1,11 @@
+export * from "./accounts";
+export * from "./constants";
+export * from "./definedTypes";
+export * from "./error";
+export * from "./events";
+export * from "./field";
+export * from "./idl";
+export * from "./instruction";
+export * from "./metadata";
+export * from "./type";
+export * from "./state";
