+++ spec/src/types/metadata.ts
@@ -1,0 +1,16 @@
+/**
+ * Metadata for IDL origination and program identification
+ */
+export type IdlMetadata = {
+    origin?: string;
+    deployments?: IdlMetadataDeployments[]
+};
+
+/**
+ * Information related to where the relevant program is
+ */
+export type IdlMetadataDeployments = {
+    chainId: string
+    address: string;
+    slot?: number
+}
