+++ spec/src/types/state.ts
@@ -1,0 +1,9 @@
+import { IdlTypeDef } from './definedTypes'
+import { IdlInstruction } from './instruction';
+
+export type IdlState = {
+    struct: IdlTypeDef;
+    methods: IdlStateMethod[];
+};
+
+export type IdlStateMethod = IdlInstruction;
