+++ spec/tests/e2e.test.ts
@@ -1,0 +1,4 @@
+import IdlSchema from '../src';
+import sample from "../examples/xnft.json";
+
+console.log(IdlSchema.safeParse(sample));
